Made Positions unsigned long long. Updated hash.c and hash.h to implement overflow checking. -SL
git-svn-id: https://gamescrafters.svn.sourceforge.net/svnroot/gamescrafters/GamesmanClassic/trunk@1429 a668be5d-ef11-cb4c-8bfd-e08cbb24f2ba

+++ gamesman/src/gamesman.h
@@ -58,7 +58,7 @@ typedef int ClientData;
 #define  TRUE           1
 #endif
 
-typedef unsigned long POSITION;      /* This used to be determined by mTempTypedef.h */
+typedef unsigned long long POSITION;      /* This used to be determined by mTempTypedef.h */
 typedef int MOVE;           /* but I decided to hardcode it for now, so as  */
 typedef int MOVES;          /* to make it a leeeetle bit easier and cleaner */
 #define POSITION_FORMAT "%ld"
